Add RestClient.major_version
