+++ lib/em-rest-client.rb
@@ -1,2 +1,3 @@
-require 'em-rest-client/rest_client'
+require 'em-rest-client/configuration'
+require 'em-rest-client/request'
 require 'em-rest-client/em_http_request'
